Introduce DiscoveryCode topic
