+++ aardvark-node/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod network;
 pub mod operation;
 pub mod document;
+pub mod topics;
