+++ aardvark-node/src/operation.rs
@@ -6,7 +6,7 @@ use p2panda_store::{LogStore, MemoryStore};
 use p2panda_stream::operation::ingest_operation;
 use serde::{Deserialize, Serialize};
 
-use crate::document::TextDocument;
+use crate::topics::TextDocument;
 
 pub(crate) type LogId = Hash;
